+++ pybillomat/clients.py
@@ -852,7 +852,10 @@ class Clients(list):
         self.per_page = int(clients_etree.attrib.get("per_page", "0"))
         self.total = int(clients_etree.attrib.get("total", "0"))
         self.page = int(clients_etree.attrib.get("page", "1"))
-        self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        try:
+            self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        except ZeroDivisionError:
+            self.pages = 0
 
         # Iterate over all clients
         for client_etree in clients_etree: