+++ pybillomat/recurring_email_receivers.py
@@ -234,7 +234,10 @@ class RecurringEmailReceivers(list):
         self.per_page = int(recurring_email_receivers_etree.attrib.get("per_page", "100"))
         self.total = int(recurring_email_receivers_etree.attrib.get("total", "0"))
         self.page = int(recurring_email_receivers_etree.attrib.get("page", "1"))
-        self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        try:
+            self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        except ZeroDivisionError:
+            self.pages = 0
 
         # Iterate over all items
         for recurring_email_receiver_etree in recurring_email_receivers_etree: