+++ pybillomat/recurring_items.py
@@ -316,7 +316,10 @@ class RecurringItems(list):
         self.per_page = int(recurring_items_etree.attrib.get("per_page", "100"))
         self.total = int(recurring_items_etree.attrib.get("total", "0"))
         self.page = int(recurring_items_etree.attrib.get("page", "1"))
-        self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        try:
+            self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        except ZeroDivisionError:
+            self.pages = 0
 
         # Iterate over all items
         for recurring_item_etree in recurring_items_etree: