+++ pybillomat/recurrings.py
@@ -641,7 +641,10 @@ class Recurrings(list):
         self.per_page = int(recurrings_etree.attrib.get("per_page", "100"))
         self.total = int(recurrings_etree.attrib.get("total", "0"))
         self.page = int(recurrings_etree.attrib.get("page", "1"))
-        self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        try:
+            self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))
+        except ZeroDivisionError:
+            self.pages = 0
 
         # Iterate over all recurrings
         for recurring_etree in recurrings_etree: